Add RSS feed generation
Support for primitive RSS feed generation has been added. At the moment,
the RSS is dinamically regenerated on every request, which may not be
the most efficient approach. Ideally, there would be a static filewhich
would be updated on demand.

Currently the RSS feed shows all items in the database, even if they
have a date that is in the past. It might be better to switch to
displaying only events, such as newly created items, deletions and
changes of entries, or a summary of entries ocurring this week or the
next day...